Rollback in GroupsDAO

+++ forest-001/src/main/java/com/kukvar/hibernate/DAO/GroupsDAO.java
@@ -1,9 +1,11 @@
 package com.kukvar.hibernate.DAO;
 
+import java.sql.SQLIntegrityConstraintViolationException;
 import java.util.List;
 
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
+import org.hibernate.Transaction;
 
 import com.kukvar.hibernate.entity.Category;
 import com.kukvar.hibernate.entity.Group;
@@ -11,7 +13,7 @@ import com.kukvar.hibernate.utils.HibernateUtil;
 
 public class GroupsDAO {
 	SessionFactory factory = HibernateUtil.getSessionFactory();
-	
+
 	public int addGroupDetails(Group group) {
 		Session session = factory.openSession();
 		session.beginTransaction();
@@ -19,8 +21,8 @@ public class GroupsDAO {
 		session.getTransaction().commit();
 		session.close();
 		return id;
-	}	
-	
+	}
+
 	public List<Group> listGroups() {
 		Session session = factory.openSession();
 		session.beginTransaction();
@@ -29,8 +31,8 @@ public class GroupsDAO {
 		session.getTransaction().commit();
 		session.close();
 		return groups;
-	}	
-	
+	}
+
 	public void updateInformation(int id, String name, String description, String nameImageFile, Category category) {
 		Session session = factory.openSession();
 		session.beginTransaction();
@@ -42,8 +44,8 @@ public class GroupsDAO {
 		group.setCategory(category);
 		session.getTransaction().commit();
 		session.close();
-	}	
-	
+	}
+
 	public Group getGroup(int id) {
 		Session session = factory.openSession();
 		session.beginTransaction();
@@ -51,28 +53,28 @@ public class GroupsDAO {
 		session.getTransaction().commit();
 		session.close();
 		return group;
-	}	
-	
+	}
+
 	public Group getGroup(String name) {
 		Session session = factory.openSession();
 		session.beginTransaction();
-		String queryString = "from classes where name = '"+name+"'";
+		String queryString = "from classes where name = '" + name + "'";
 		Group group;
 		try {
 			group = (Group) session.createQuery(queryString).getSingleResult();
 		} catch (Exception e) {
-			// TODO: handle exception  GroupsDAO
+			// TODO: handle exception GroupsDAO
 			return null;
 		}
 		session.getTransaction().commit();
 		session.close();
 		return group;
 	}
-	
+
 	public boolean isExisted(String name) {
 		Session session = factory.openSession();
 		session.beginTransaction();
-		String queryString = "from classes where name = '"+name+"'";
+		String queryString = "from classes where name = '" + name + "'";
 		@SuppressWarnings("unchecked")
 		List<Group> groups = session.createQuery(queryString).getResultList();
 		session.getTransaction().commit();
@@ -82,16 +84,36 @@ public class GroupsDAO {
 		} else {
 			return true;
 		}
-		
-	}	
-	
-	public void deleteGroup(int id) {
-		Session session = factory.openSession();
-		session.beginTransaction();		
-		Group group = session.get(Group.class, id);
-		session.delete(group);
-		session.getTransaction().commit();
-		session.close();
+
 	}
-	
+
+	public void deleteGroup(int id) throws SQLIntegrityConstraintViolationException {
+		Session session = factory.openSession();
+		Transaction txn = session.getTransaction();
+		try {
+			txn.begin();
+			Group group = session.get(Group.class, id);
+			if (group != null) {
+				session.delete(group);
+			}
+			txn.commit();
+		} catch (Exception e) {
+			if (txn != null) {
+				txn.rollback();
+			}
+			if (e.getCause() != null && e.getCause().getCause() instanceof SQLIntegrityConstraintViolationException) {
+				SQLIntegrityConstraintViolationException sql_violation_exception = (SQLIntegrityConstraintViolationException) e
+						.getCause().getCause();
+				throw sql_violation_exception;
+			} else {
+				e.printStackTrace();
+				throw e;
+			}
+		} finally {
+			if (session != null) {
+				session.close();
+			}
+		}
+	}
+
 }
